update timer sample and fix default actor
