+++ sample/src/main/java/com/arttttt/simplemvi/sample/timer/TimerContent.kt
@@ -14,6 +14,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
+import com.arttttt.simplemvi.sample.timer.store.TimerStore
 import com.arttttt.simplemvi.utils.plus
 
 @Composable
