+++ sample/src/main/java/com/arttttt/simplemvi/sample/timer/TimerViewModel.kt
@@ -1,6 +1,7 @@
 package com.arttttt.simplemvi.sample.timer
 
 import androidx.lifecycle.ViewModel
+import com.arttttt.simplemvi.sample.timer.store.TimerStore
 import com.arttttt.simplemvi.viewmodel.attachStore
 import kotlinx.coroutines.Dispatchers
